% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stistko.R
\name{get_stistko_ciselnik}
\alias{get_stistko_ciselnik}
\title{Get STISTKO Ciselnik (df from code)}
\usage{
get_stistko_ciselnik(code, dest_dir = NULL)
}
\arguments{
\item{code}{A character string representing the code of the codelist.}
}
\value{
A data frame containing the processed data from the ciselnik.
}
\description{
Reads and processes the HTML file of a STISTKO ciselnik based on a code
}
\examples{
get_stistko_ciselnik("BASO")
}
