% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_directory.R
\name{response_to_quasixls}
\alias{response_to_quasixls}
\title{Turn a httr response created by \code{vz_get_directory_responses()} into and XLS file}
\usage{
response_to_quasixls(response, tempfile = TRUE)
}
\arguments{
\item{response}{a httr respons returned by \code{vz_get_directory_responses()}}

\item{tempfile}{whether to write into a tempfile (TRUE, the default), or locally}
}
\value{
character of length 1: path to XLS file
}
\description{
Turn a httr response created by \code{vz_get_directory_responses()} into and XLS file
}
