% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsezved-package.R
\docType{package}
\name{vsezved-package}
\alias{vsezved}
\alias{vsezved-package}
\title{vsezved: Get data on Czech schools from <https://uiv.cz/> and <https://data.msmt.cz/>}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{http://petrbouchal.xyz/vsezved}
  \item \url{https://github.com/petrbouchal/vsezved}
  \item Report bugs at \url{https://github.com/petrbouchal/vsezved/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
