% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stistko.R
\name{vz_download_codelist}
\alias{vz_download_codelist}
\title{Download STISTKO Ciselnik}
\usage{
vz_download_codelist(url, dest_dir = NULL)
}
\arguments{
\item{url}{A character string representing the URL to download the HTML page from.}

\item{dest_dir}{A character string specifying the destination directory. Defaults to tempdir().}
}
\value{
A character string containing the path to the downloaded HTML file.
}
\description{
Downloads the HTML page from the given URL.
}
\examples{
vz_download_codelist("http://stistko.uiv.cz/katalog/ciselnik11x.asp?idc=BASO&aap=on")
}
