% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stistko.R
\name{vz_get_codelist}
\alias{vz_get_codelist}
\title{Get Stistko Ciselnik (df from code)}
\usage{
vz_get_codelist(code, dest_dir = NULL)
}
\arguments{
\item{code}{A character string representing the code of the codelist.}
}
\value{
A data frame containing the processed data from the ciselnik.
}
\description{
Reads and processes the HTML file of a Stistko ciselnik based on a code
}
\examples{
vz_get_codelist("BASO")
}
