% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stistko.R
\name{vz_get_codelist_url}
\alias{vz_get_codelist_url}
\title{Get URL for STISTKO Ciselnik}
\usage{
vz_get_codelist_url(code)
}
\arguments{
\item{code}{A character string representing the ciselnik code.}
}
\value{
A character string containing the URL for the specified ciselnik code.
}
\description{
Constructs the URL for the specified STISTKO codelist code.
}
\examples{
get_stistko_ciselnik_url("BASO")
}
