% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_directory.R
\name{vz_get_directory_responses}
\alias{vz_get_directory_responses}
\title{Get school directory responses}
\usage{
vz_get_directory_responses(
  tables = c("addresses", "schools", "locations", "specialisations"),
  ...
)
}
\arguments{
\item{tables}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
