% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_register.R
\name{vz_get_register}
\alias{vz_get_register}
\title{FUNCTION_TITLE}
\usage{
vz_get_register(
  dataset_id = "rejstrik-skol-a-skolskych-zarizeni-cela-cr",
  tables = "organisations",
  keep_file = T
)
}
\arguments{
\item{dataset_id}{DESCRIPTION.}

\item{tables}{DESCRIPTION.}

\item{keep_file}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
