% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_register.R
\name{vz_get_register_xml}
\alias{vz_get_register_xml}
\title{Download (XML) file of register}
\usage{
vz_get_register_xml(
  url = NULL,
  nuts3_kod = NULL,
  write_file = F,
  dest_dir = getwd()
)
}
\arguments{
\item{url}{URL; if left to NULL, will use internal default}

\item{nuts3_kod}{used to point to per-region datasets; if left unset, defaults to state-wide data}

\item{write_file}{Whether to keep the downloaded XML file.
Currently only writing to the working directory is supported.}

\item{dest_dir}{Where to write the resulting XML}
}
\value{
Path to downloaded (XML) file.
}
\description{
Uses CKAN to find the correct URL in the education ministry's \href{https://data.msmt.cz/}{open data catalogue} and retrieve the file.
}
