% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_directory.R
\name{vz_get_search_form}
\alias{vz_get_search_form}
\title{Get search form for school directory}
\usage{
vz_get_search_form(search_page = NULL)
}
\arguments{
\item{search_page}{search page session as returned by \code{vz_get_search_page()}}
}
\value{
An rvest_form object to be passed on to \code{vz_get_directory_responses()}.
}
\description{
Get search form for school directory
}
