% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared.R
\name{vz_get_xml_url}
\alias{vz_get_xml_url}
\title{Get URL of file in MSMT data store}
\usage{
vz_get_xml_url(nuts3_kod = NULL, base_url = NULL)
}
\arguments{
\item{nuts3_kod}{NUTS code for region, e.g. CZ0100 for Prague. Leave as NULL for whole-country school register.}

\item{base_url}{Base URL. Leave as NULL for MSMT data store URL.}
}
\value{
a URL, character of length 1
}
\description{
Currently assumes we are getting register XML data
}
