% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stistko.R
\name{vz_grab_codelist}
\alias{vz_grab_codelist}
\title{Get STISTKO Ciselnik}
\usage{
vz_grab_codelist(code, dest_dir = NULL)
}
\arguments{
\item{code}{A character string representing the ciselnik code.}

\item{dest_dir}{A character string specifying the destination directory. Defaults to tempdir().}
}
\value{
A character string containing the path to the downloaded HTML file.
}
\description{
Downloads the HTML page for the specified STISTKO codelist code.
}
\examples{
vz_grab_codelist("BASO")
}
