% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_register.R
\name{vz_list_registers}
\alias{vz_list_registers}
\title{List packages (registers) available from CKAN catalogue}
\usage{
vz_list_registers(url = NULL)
}
\arguments{
\item{url}{If left to NULL, searches MSMT catalogue at <data.msmt.cz>}
}
\value{
a character vector of IDs
}
\description{
By default, searches MSMT
}
