% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_directory.R
\name{vz_load_directory}
\alias{vz_load_directory}
\title{Load directory XLS file}
\usage{
vz_load_directory(path)
}
\arguments{
\item{path}{Path to .xls file retrieved by \code{vz_write_directory_quasixls}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Read and clean up quasi-XLSX file retrieved by \code{vz_write_directory_quasixls}
}
