% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stistko.R
\name{vz_read_codelist}
\alias{vz_read_codelist}
\title{Read STISTKO Ciselnik}
\usage{
vz_read_codelist(path)
}
\arguments{
\item{path}{A character string representing the path to the HTML file.}
}
\value{
A data frame containing the processed data from the ciselnik.
}
\description{
Reads and processes the HTML file of a STISTKO ciselnik.
}
