% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_directory.R
\name{vz_write_directory_quasixls}
\alias{vz_write_directory_quasixls}
\title{Turn a httr response created by \code{vz_get_directory_responses()} into and XLS file}
\usage{
vz_write_directory_quasixls(response, write_file = FALSE, dest_dir = getwd())
}
\arguments{
\item{response}{a httr respons returned by \code{vz_get_directory_responses()}}

\item{dest_dir}{Directory in which to write XLS files. Defaults to working directory.}
}
\value{
character of length 1: path to XLS file
}
\description{
Turn a httr response created by \code{vz_get_directory_responses()} into and XLS file
}
