% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vsezved-package.R
\docType{package}
\name{vsezved-package}
\alias{vsezved}
\alias{vsezved-package}
\title{vsezved: Get data on Czech schools from <https://stistko.uiv.cz/registr/> and <https://data.msmt.cz/>}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Get access to data on Czech schools: the register open data provided by the Ministry of Education at \url{https://data.msmt.cz/} and non-open web database at \url{http://stistko.uiv.cz/registr/} This is mostly organisatoinal data on primary and secondary schools.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://petrbouchal.xyz/vsezved}
  \item \url{https://github.com/petrbouchal/vsezved}
  \item Report bugs at \url{https://github.com/petrbouchal/vsezved/issues}
}

}
\author{
\strong{Maintainer}: Petr Bouchal \email{pbouchal@gmail.com} (\href{https://orcid.org/0000-0002-0471-716X}{ORCID})

}
\keyword{internal}
