% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_register.R
\name{vz_get_register}
\alias{vz_get_register}
\title{Download and read school register}
\usage{
vz_get_register(
  nuts3_kod = NULL,
  url = NULL,
  tables = c("organisations", "schools", "locations", "specialisations"),
  write_file = TRUE,
  dest_dir = getwd()
)
}
\arguments{
\item{nuts3_kod}{used to point to per-region datasets; if left unset, defaults to state-wide data}

\item{url}{URL; if left to NULL, will use internal default}

\item{tables}{Which tables to return. Can be one or more of "organisations",
"schools", "locations" or "specialisations" (specialisations not yet available via the package).}

\item{write_file}{Whether to keep the downloaded XML file.
Currently only writing to the working directory is supported.}

\item{dest_dir}{Where to write the resulting XML}
}
\value{
a \link[tibble:tibble-package]{tibble} or list of tibbles if multiple
table names are passed to \code{tables}.
}
\description{
This is the high-level function for getting data from the online XML export
of the school register. It downloads the file (whole country by default) and
turns it into a tibble, cleaning up names and dropping some uninteresting
columns (this may change as the package matures.)
}
