% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/school_register.R
\name{vz_load_register}
\alias{vz_load_register}
\title{FUNCTION_TITLE}
\usage{
vz_load_register(dl_path, tables = "organisations")
}
\arguments{
\item{dl_path}{DESCRIPTION.}

\item{tables}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
